+++ src/includes/globals/.load.php
@@ -9,6 +9,6 @@ declare (strict_types = 1);
 namespace WebSharks\WpSharks\Core\Globals;
 
 if (!defined('WPINC')) {
-    exit('Do NOT access this file directly: '.basename(__FILE__));
+    exit('Do NOT access this file directly.');
 }
 require_once __DIR__.'/wp-version.php';