+++ src/includes/globals/wp-version.php
@@ -11,7 +11,7 @@ declare (strict_types = 1);
 namespace WebSharks\WpSharks\Core\Globals;
 
 if (!defined('WPINC')) {
-    exit('Do NOT access this file directly: '.basename(__FILE__));
+    exit('Do NOT access this file directly.');
 }
 if (!defined('WP_VERSION')) {
     define('WP_VERSION', $GLOBALS['wp_version']);