+++ src/includes/stub.php
@@ -9,6 +9,6 @@ declare (strict_types = 1);
 namespace WebSharks\WpSharks\Core;
 
 if (!defined('WPINC')) {
-    exit('Do NOT access this file directly: '.basename(__FILE__));
+    exit('Do NOT access this file directly.');
 }
 require_once dirname(__FILE__, 2).'/vendor/autoload.php';