+++ src/includes/templates/s-core/admin/menu-pages/post-meta-box/template.php
@@ -20,6 +20,9 @@ use WebSharks\Core\WpSharksCore\Traits as CoreTraits;
 use function assert as debug;
 use function get_defined_vars as vars;
 
+if (!defined('WPINC')) {
+    exit('Do NOT access this file directly.');
+}
 ?>
 <div class="<?= esc_attr($cfg->class); ?>">
     <div class="-container">