+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Correcting typo in dependency notice. `resolove` should be `resolve`.
 - Bug fix. Broken 'Settings' link for some plugins running on the WP Sharks Core. See [Issue #5](https://github.com/websharks/wp-sharks-core/issues/5).
 - Enhancing WooCommerce conditionals in `Wp` base class.
+- Enhancing security by removing `basename(__FILE__)` from notices.
 
 ## v161026.47592